cleanup for public mirroring of branches
This cleansup master to provide direction to users that they can find
proper examples in the `rel/v` branches. Removes all the information
from the master branch so that during public mirrors the copy over
is off from a relatively clean branch.

This mimics the behavior of our other public repository `datawire/edge-stack`.

Signed-off-by: Lance Austin <laustin@datawire.io>